+++ Commons/TransactionApi/src/main/java/com/mfino/transactionapi/handlers/subscriber/impl/SubscriberRegistrationThroughWebHandlerImpl.java
@@ -238,7 +238,7 @@ public class SubscriberRegistrationThroughWebHandlerImpl extends FIXMessageHandl
 		Subscriber subscriber = subscriberMDN.getSubscriber();
         SubscriberGroupDao subscriberGroupDao=new DAOFactory().getInstance().getSubscriberGroupDao();
 		
-		Set<SubscriberGroups> subscriberGroups = (Set<SubscriberGroups>) subscriberGroupDao.getAllBySubscriberID(subscriber.getId());
+		Set<SubscriberGroups> subscriberGroups = (Set<SubscriberGroups>) subscriberGroupDao.getAllBySubscriberID(BigDecimal.valueOf(subscriber.getId()));
 		if(subscriberGroups != null && !subscriberGroups.isEmpty())
 		{
 			SubscriberGroups subscriberGroup = subscriberGroups.iterator().next();
@@ -284,7 +284,7 @@ public class SubscriberRegistrationThroughWebHandlerImpl extends FIXMessageHandl
 		Long groupID = null;
         SubscriberGroupDao subscriberGroupDao=new DAOFactory().getInstance().getSubscriberGroupDao();
 		
-		Set<SubscriberGroups> subscriberGroups = (Set<SubscriberGroups>) subscriberGroupDao.getAllBySubscriberID(subscriber.getId());
+		Set<SubscriberGroups> subscriberGroups = (Set<SubscriberGroups>) subscriberGroupDao.getAllBySubscriberID(BigDecimal.valueOf(subscriber.getId()));
 		if(subscriberGroups != null && !subscriberGroups.isEmpty())
 		{
 			SubscriberGroups subscriberGroup = subscriberGroups.iterator().next();
